+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Mcoin (Bcoin for Monacoin) Release Notes & Changelog
+
+## v1.0.0-beta.15
+
+### Notable Changes
+
+- Convert Monacoin.
+
 # Bcoin Release Notes & Changelog
 
 ## v1.0.0-beta.14